requirements.txt
acres==0.1.0
aiobotocore==2.15.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioitertools==0.12.0
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
apache-airflow==2.10.4
apache-airflow-providers-celery==3.8.5
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.2
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-google==12.0.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.1
argcomplete==3.5.3
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
botocore==1.35.36
build==1.2.2.post1
cachelib==0.9.0
cachetools==5.5.0
captcha-solver==0.1.5
cattrs==24.1.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
configobj==5.0.9
configparser==7.1.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
db-dtypes==1.4.0
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.9
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
durationpy==0.9
email_validator==2.2.0
etelemetry==0.3.1
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
filetype==1.2.0
fitz==0.0.1.dev2
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
flower==2.0.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec
gcloud-aio-auth==5.3.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.3.0
gcsfs
google-ads==25.1.0
google-analytics-admin==0.23.3
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.78.0
google-cloud-alloydb==0.4.1
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-automl==2.15.0
google-cloud-batch==0.17.33
google-cloud-bigquery==3.29.0
google-cloud-bigquery-datatransfer==3.18.0
google-cloud-bigtable==2.28.1
google-cloud-build==3.29.0
google-cloud-compute==1.23.0
google-cloud-container==2.55.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.24.1
google-cloud-dataflow-client==0.8.15
google-cloud-dataform==0.5.14
google-cloud-dataplex==2.6.0
google-cloud-dataproc==5.16.0
google-cloud-dataproc-metastore==1.17.0
google-cloud-dlp==3.26.0
google-cloud-kms==3.2.2
google-cloud-language==2.16.0
google-cloud-logging==3.11.3
google-cloud-memcache==1.11.0
google-cloud-monitoring==2.26.0
google-cloud-orchestration-airflow==1.16.1
google-cloud-os-login==2.16.0
google-cloud-pubsub==2.27.3
google-cloud-redis==2.17.0
google-cloud-resource-manager==1.14.0
google-cloud-run==0.10.14
google-cloud-secret-manager==2.22.1
google-cloud-spanner==3.51.0
google-cloud-speech==2.30.0
google-cloud-storage==2.19.0
google-cloud-storage-transfer==1.15.0
google-cloud-tasks==2.18.0
google-cloud-texttospeech==2.24.0
google-cloud-translate==3.19.0
google-cloud-videointelligence==2.15.0
google-cloud-vision==3.9.0
google-cloud-workflows==1.16.0
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpc-google-iam-v1==0.14.0
grpc-interceptor==0.15.4
grpcio==1.70.0
grpcio-gcp==0.2.2
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
humanize==4.11.0
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
json-merge-patch==0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kombu==5.4.2
kubernetes==31.0.0
lazy-object-proxy==1.10.0
limits==4.0.1
linkify-it-py==2.0.3
llama-cloud==0.1.6
llama-index==0.12.5
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.5
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.10
llama-index-multi-modal-llms-openai==0.4.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.17
lockfile==0.12.2
looker-sdk==25.0.0
looseversion==1.3.0
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mmh3==5.0.1
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.26.0
neomodel==5.4.1
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.57.2
openparse==0.7.0
ordered-set==4.1.0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandas-gbq==0.26.1
pathlib==1.0.1
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pendulum==3.0.0
pillow==11.0.0
playwright==1.43.0
pluggy==1.5.0
posthog==3.7.3
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
prov==2.0.1
psutil==6.1.1
puremagic==1.28
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydata-google-auth==1.9.1
pydot==3.0.2
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.24.14
pyOpenSSL==25.0.0
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==8.2.2
pytest-mock==3.14.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.22.3
rsa==4.9
s3fs==2024.10.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setproctitle==1.3.4
setuptools==70.0.0
shapely==2.0.6
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
slugify==0.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.1
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.8.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.40.0
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tika==2.6.0
tiktoken==0.8.0
time-machine==2.16.0
tokenizers==0.20.3
torch==2.5.1
tornado==6.4.2
tqdm==4.66.6
traits==6.4.3
transformers==4.46.3
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wirerope==1.0.0
wordcloud==1.9.4
wrapt==1.17.0
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
